requirements.txt
Django==2.2.19
djangorestframework==3.12.2
psycopg2==2.8.5

# CORS
django-cors-headers==3.5.0

djangorestframework-jwt==1.11.0
gunicorn==20.0.4
whitenoise==5.2.0
requests==2.25.1